boom boom

+++ courses/dl2/imdb_scripts/imdb_small_experiment.py
@@ -17,9 +17,10 @@ ORIG_SMALL_DATA_DIR = Path('/home/paperspace/text-augmentation/imdb_1k3k/')
 
 
 def make_small_ds(src_path, dest_path, n_train, n_test=3000):
-    shutil.rmtree(dest_path)
+
     if dest_path is None:
         dest_path = Path(f'/home/paperspace/imdb_{int(n_train/1000)}k_{int(n_test/1000)}k/')
+        if dest_path.exist(): shutil.rmtree(dest_path)
         dest_path.mkdir(exist_ok=True)
         print(dest_path)
